% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample_phyloseq.R
\name{subsample_phyloseq}
\alias{subsample_phyloseq}
\title{subsample_phyloseq}
\usage{
subsample_phyloseq(ps_object, repeats = 100, depth = 50000, n_cores = 1, ...)
}
\arguments{
\item{ps_object}{An existing phyloseq object to repeatedly subsample.}

\item{repeats}{Number of repeats (defaults to n=100).}

\item{depth}{Library size that each sample will be subsampled to. Sampling is performed without replacement and a value required to encourage reporting of the depth (defaults to 50k).}

\item{n_cores}{Number of cores to use for parallel computations (default n=1).}

\item{...}{}
}
\value{
A list with the following elements:

ps_object: The input ps_object (i.e., unsubsampled object)

repeats: The number of repeated subsamples

subsample_depth: The library size selected for subsampling

adiv: Alpha-diversity metrics averaged across subsamples

bray_curtis: Bray-Curtis dissimilarity averaged across subsamples

jaccard: Jaccard index averaged across subsamples

horn: Horn index averaged across subsamples

subsampled_list: returns a list of the subsampled phyloseq objects
}
\description{
This function performs repeated subsampling of an existing phyloseq object and returns the average across subsamples for common alpha- and beta-diversity metrics. Subsampling is performed without replacement and parallel computations supported. A list of subsampled phyloseq objects is returned allowing for additional metrics to be computed and averaged across the subsamples as required.
}
